GNL final

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: aarias-d <aarias-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/13 16:24:42 by aarias-d          #+#    #+#             */
-/*   Updated: 2025/06/24 18:23:31 by aarias-d         ###   ########.fr       */
+/*   Updated: 2025/07/15 17:04:17 by aarias-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,10 +23,10 @@
 # include <unistd.h>
 
 char	*get_next_line(int fd);
-void    ft_free(char* pointer);
+void	ft_free(char *pointer);
 char	*ft_strchr(char *s, int c);
 char	*ft_strjoin(char *s1, char *s2);
-size_t	ft_strlen(char *str);
+int		ft_strlen(char *str);
 char	*ft_strdup(char *src);
 
 #endif
